fix: Data type conversion issue

+++ Commander/BSC/BSC_SubscriptionCommander.sol
@@ -96,7 +96,7 @@ contract BSC_SubscriptionCommander is Subscription, Commander {
             ) {
                 if (refundCountForDisable >= _percentValue(packInfo.total - quantity, 60)) {
                     disabledPack = true;
-                    disabledTime = block.timestamp;
+                    disabledTime = uint32(block.timestamp);
                 }
 
                 refundCountForDisable++;
